pyproject.toml
[project]
name = "lithium"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "channels>=4.2.2",
    "coverage>=7.8.0",
    "crispy-bootstrap5>=2025.4",
    "debugpy>=1.8.14",
    "dj-database-url>=2.3.0",
    "django>=5.2",
    "django-admin-env-notice>=1.0.1",
    "django-allauth>=65.7.0",
    "django-cors-headers>=4.7.0",
    "django-crispy-forms>=2.4",
    "django-dbbackup>=4.2.1",
    "django-debug-toolbar>=5.2.0",
    "django-environ>=0.12.0",
    "django-extensions>=4.1",
    "django-filter>=25.1",
    "django-import-export>=4.3.7",
    "django-oauth-toolkit>=3.0.1",
    "django-phonenumber-field>=8.1.0",
    "django-polymorphic>=3.1.0",
    "django-post-office>=3.9.1",
    "django-q2>=1.8.0",
    "django-redis>=5.4.0",
    "django-ses>=4.4.0",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.0",
    "djangorestframework-datatables>=0.7.2",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-spectacular>=0.28.0",
    "factory-boy>=3.3.3",
    "faker>=37.1.0",
    "freezegun>=1.5.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "langchain>=0.3.25",
    "markdown>=3.8",
    "openai>=1.77.0",
    "pandas>=2.2.3",
    "phonenumbers>=9.0.4",
    "plotly>=6.0.1",
    "pre-commit>=4.2.0",
    "psycopg>=3.2.7",
    "pytest>=8.3.5",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "qrcode>=8.0",
    "pillow>=10.0.0",
    "ruff>=0.11.8",
    "setuptools>=80.1.0",
    "weasyprint>=65.1",
    "whitenoise>=6.9.0",
]

[dependency-groups]
dev = [
  "typing_extensions ~=4.12",
]
